Rename file main to person
